research/requirements.txt
# Core ML and Data Science
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Model Export and Conversion
onnx>=1.14.0
onnxruntime>=1.15.0
tensorflow>=2.13.0

# Data Loading and Processing
datasets>=2.14.0
huggingface-hub>=0.16.0
tqdm>=4.65.0

# Visualization and Analysis
plotly>=5.15.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.0.0